Merged PR 621628: Publish app.config when binding redirects are specified
This PR fixes the BuildXL SDK to produce 'AssemblyName.dll.config' file even when the `appConfig` property was not provided in `BuildXLSdk.test` call.

I.e. now it should be sufficient to have the following code:

```
@@public
export const dll = BuildXLSdk.test({
    assemblyName: "MyTest",
    // No appConfig: f`App.config`,
    assemblyBindingRedirects: BuildXLSdk.cacheBindingRedirects()});
```

Before this change, it was necessary to have an empty `App.config` file and pass it to the `BuildXLSdk.test` method invocation.