Merged PR 621515: Store the hash->buildManifestHash mapping in historic metadata cache
This PR makes a simple change to store the hash-hash mapping in historic metadata cache, rather than in actual cache. The mapping is persistent across the builds; the data from workers is added to the HMC on the orchestrator (the orchestrator is the one that saves HMC to cache at the end of a build).

My very unscientific test (20 runs of 280k pip builds - 5 clean builds with manifest on/off, 5 cached builds with manifest on/off) showed no obvious perf degradation. Obviously, more tests are needed.

The only concern is historic metadata cache miss (i.e., not loading rocksdb from cache at the beginning of a build). In the last 10 days this happened to 7% of builds (could not paste Kusto, so adding a pic):

![image.png](https://dev.azure.com/mseng/9ed2c125-1cd5-4a17-886b-9d267f3a5fab/_apis/git/repositories/50d331c7-ea65-45eb-833f-0303c6c2387e/pullRequests/621515/attachments/image.png)

Interestingly, the miss rate really depends on a queue:
![image (2).png](https://dev.azure.com/mseng/9ed2c125-1cd5-4a17-886b-9d267f3a5fab/_apis/git/repositories/50d331c7-ea65-45eb-833f-0303c6c2387e/pullRequests/621515/attachments/image%20%282%29.png)

Related work items: #1854580