Merged PR 591177: Extract 'Error' object to make Result API clearer
The main goal for this PR to make the semantics of  result types clearer and make it easier to understand when the result is successful and when its not.

Now every result type has 'Error' property and the result is successful only when `Error != null` and `IsCancelled` is false. Some result types still override `Succeeded` property but only to add `MemberNotNullWhen` attribute.

If the class needs to control whether the result is successful or not it overrides `Error` property instead.

This PR also fixes a few bugs, for instance, before this change `new Result<string>(string.Empty){IsCancelled}` was considered a successful result.