Merged PR 623772: High priority for light process pips in Scheduler
Process pips and IPC pips stay in the same dispatcher when choosing a worker. We dequeue the highest priority item and try to choose a worker. If the highest priority pip cannot acquire a slot, the ChooseWorker dispatcher is blocked and all other pips wait in the queue. We fix this issue for IPC pips by specifying very high priority. We do the same fix for light process pips as well.

Related work items: #1862378