Merged PR 623531: Disambiguate ResultBase constructors
One of the changes in [Extract 'Error' object to make Result API clear](https://dev.azure.com/mseng/Domino/_git/BuildXL.Internal/pullrequest/591177) PR was not good: I changed the signature for the successful case in `ResultBase(string successDiagnostics)`. The issue is that before `Result(string)` was used for creating an unsuccessful results.

So this PR changes the successful constructor signature by removing the string argument.