Merged PR 624456: Expose 'RemoveFromTracker' in LocalContentServer
This PR exposes 'RemoveFromTracker' API in `LocalContentServer` that will allow AnyBuild to "unregister" the content at shutdown without any RPC calls.

And the PR also changes the result for `RemoveFromTracker` API to return `BoolResult` instead of `Result<long>`. This API used to return the number of un-registered entries, but this no longer true for a long time (once LLS was implemented).

This is part of the fix for #1840749