Merged PR 624645: Annotate ILogger with StringFormatMethodAttribute
The cache codebase does not use `ILogger` interface from the cache layer that takes formatting argument too much, but AnyBuild does use it a bit more.

This PR adds `StringFormatMethod` attribute annotations for the methods that takes extra params to improve the IDE support and to allow the analyzers (ErrorProne.NET in particular) to detect the incorrect usages of such methods.