Merged PR 625367: Fix 'Missing method' error when calling 'CreateVssCredentialsAsync' in .net core
TIL how the CLR validates the correctness of the method:
Consider the following method:

```csharp
public void FooBar(bool b)
{
  if (b)
  {
      MethodThatDoesNotExist();
  }
}
```

`MethodThatDoesNotExist` is a method that exists for .net472 and doesn't exist for .net core and we try to run `FooBar(false)` on .net core. One may think that because the call to `MethodThatDoesNotExist` is not reachable, the call `FooBar(false)` will be successful.

But this is not the case! The CLR will check that the method body is correct before calling the method. So the call will fail.

This is exactly what was happening when we tried running `CreateVssCredentialsAsync` under .NET CORE. Even though the user name was specified, the call to `CreateVssCredentialsAsync` was failing with 'missing method' exception even though we never called `_helper.GetCredentialsAsync` (the method that is only available for full framework.