Merged PR 625850: Adding a new dispatcher to choose a worker for light pips
We want to have a separate dispatcher when choosing a worker for light pips. They have the highest priority due to well-known reasons, but when we do not choose a worker for them, they will block the normal process pips. It was not an issue because light dispatcher limit is around 1000. That's why, we almost always choose a worker for the light pips. However, Office noticed that this is too much for light process pips. They will decrease the limit of light dispatcher via /maxLight:100.