Merged PR 627100: ColdStorage Consistent Hashing
Use **consistent hashing** to distribute the ColdStorage data across the cluster (to avoid using the content tracking system).

**Work items:**

#1868357

Mainly:
- LocalLocationStore.cs
- ColdStorage.cs

#1869911

Mainly:
- ColdStorage.cs

#1870282

Mainly:
- GrpcContentServer.cs

#1870281

Mainly:
- Workflows.cs
- ColdStorage.cs

### About Consistent Hashing:

Consistent hashing is a strategy for dividing up keys/data between multiple machines.

How do we use it? We create a virtual ring with the active cache servers. When we need to save or search something we use the closest nodes to the content hash in the ring. This way, the only thing that each node has to know is how the ring is built. That is why we avoid that the content tracking system has to handle a lot of information.

![cons-hash.png](https://dev.azure.com/mseng/9ed2c125-1cd5-4a17-886b-9d267f3a5fab/_apis/git/repositories/50d331c7-ea65-45eb-833f-0303c6c2387e/pullRequests/627100/attachments/cons-hash.png)