Merged PR 626392: Add DScript support for expanding env variables in strings and paths
Add explicit DScript methods to expand environment variables in strings and paths. Observe the expansion needs to go through the engine to guarantee sound incrementality.
This functionality is in principle needed for paths only, but adding it for arbitrary strings as well since in some cases it is not possible to get a valid path to begin with (e.g. p`%COMSPEC` is not a valid path since it is not rooted before expansion happens).

Related work items: #1860047