Merged PR 628398: Download resolver schedules pips
With this change the download resolver schedules pips that can download and extract files (instead of doing those actions at evaluation time).
Each entry in the resolver conf that needs extraction will result in two pips: one that downloads the file, followed by another one that performs the extraction.
The only visible difference to the user wrt the former behavior is that the result of the extraction is now an opaque directory (as opposed to a partial seal one). A couple SDKs are updated accordingly to accommodate the change.

Related work items: #1865781