Merged PR 629602: Attempt to reset ownership and ACL when cache fails to set ACL on ingress
Expands the ownership and ACL setting strategies used when deleting files to also grant access to the Users group. Then employs this strategy in a few areas that need to operate on file/directories produced in the build.

The motivation for this is to deal with some pips that are producing directories with ACLs set. For those tools, the ACLs are only relevant during the time pips are executing and do not need to exist when files are retrieved from cache or drop.

Related work items: #1866269, #1871721, #1876177