Merged PR 631474: Add AppRoot to DropDaemon's explicitly forwarded environment variables
EsrpManifestSign.exe needs this variable to be correctly set - let's not rely on it being a global passthrough, rather, forward it explicitly.

This change seems to fix our issues with build manifest signing in the problematic queues. In particular the Office LTSB queues were not setting AppRoot as a passthrough.