Merged PR 631470: Correct some potential race conditions around launching takeown
A prior changed switched to async output stream handing. I suspect that as introducing a bug that is causing  process output streams to get into file paths and other funny locations. I also noticed that we were not previously ensuring the process had exited and weren't correctly closing the process handle. Now that we are calling this code more often, theses issues are more noticeable