Merged PR 630776: Fix a bunch of tests that believe they are making cache hit assertions when they actually are not
The pattern `RunScheduler().AssertCacheMiss/Hit()` is used throughout, but the callers should specify the pip ids upon which the assertion is made: calling it without pip ids amounts to no assertions at all rather than a general "something was a cache miss / everything was a cache hit".

There are a number of places where this is happening. This PR fixes the tests and add assertions to prevent this error in the future. Also add an assertion to avoid running with an empty scheduler, which probably mean the test author forgot to schedule the pip in the test.