Merged PR 632104: Only consider stale share opaques in minimal + alien enumeration mode when lazy scrubbing is enabled
Only consider stale share opaques in minimal + alien enumeration mode when lazy scrubbing is enabled. The result of profiling shows that checking for shared opaque outputs is expensive, so make sure we only do that when needed. The result is an order of magnitude faster in overall time spent on fingerprinting enumerations.