Merged PR 632647: Improve compatibility with EnforceSomeTypeSanity linter policy
We broke OSGTools because of a violation to EnforceSomeTypeSanity which we could have caught if we had this enabled in our repo. So I thought it was prudent to enable it. As I got into the changes I realized there are a few violations in netfx.dsc that are difficult to get around without declaring annoying fragile types. But I figured it was worth committing the various cleanup I had done up until that point.