Merged PR 632671: Honor setting for max number of pips to perform cache miss analysis on
Somewhere along the way we broke honoring the setting for max number of pipe to run cache miss analysis on. The effective limit with the default settings prior to this change was 1 million pips. That made for enormous cache miss analysis markdown files and is also causing cache miss analysis data to be very short lived in telemetry

Related work items: #1883339