Merged PR 634231: Bump up StringTable's Overflow Buffer count by 2
Office's mularchy queue is hitting a crash with 3 overflow buffers not being large enough. Increase it by 2.

Outside of mularchy, the largest StringTable in the last week was:
1888374602

If we ignore the added capacity allotted by the large string table and this overflow, decreasing the number of byte buffers decreases the capacity from:
4286578688
to:
4282384384
(2044 byte buffers * 2^21 bytes per buffer -> 2042 byte buffers)

So that's still quite a bit lower than non-mularchy queues would ever have to worry about. And that's omitting the large string table which increases capacity. In short this change shouldn't cause any issues for non-mularchy queues and should give mularchy queues enough headroom to complete.

Related work items: #1882961