Merged PR 633929: Differentiate between an individual DropConfig and global DropServiceConfig
With the new multiple-drop-single-service DropDaemon it doesn't make sense to attach specific drop configurations to the overarching service. The configuration we're picking up from `DropConfig.json` in most builds includes drop name/service endpoint, which only make sense for an individual drop, and also "global" build information (for SBOM purposes) makes sense to have as a global drop service configuration.

Changes on the build runner will split these into two files, one of which (say "`DropServiceConfig.json`") will hold the global information and which we will always want to expose. This PR mirrors this change on the BuildXL side.

In fact, currently there is no way to create a build-manifest-enabled drop through API alone, because there is no way to pass this build session information from the API (not even with explicit knowledge of the BSI location, etc., which is known by the build runner and spit into the `DropConfig.json` file): these values can only get through via this `dropServiceConfigFile`, and it seems like we're assuming that this will always be passed as part of the service pip configuration. This seems to have been the case until WDG started with their multidrop approach (they are not using this file, instead, launching the service with explicit API argument).

Related work items: #1883828