Merged PR 634656: Demote cache consistency warnings and soften messaging
Early versions of BuildXL had the expectation of a "consistent" cache. That meant that if there was a pathset + strong fingerprint pair, there should be an entry in the cache describing the list of content hashes associated with that strong fingerprint. And there should also be the serialized metadata buildxl needs to replay an item from cache. The idea was that we didn't want to waste time processing a pathset/strongFingerprint if that would eventually lead to a dead end.

The cache implementation has evolved quite a bit since those early days and these scenarios are not totally unexpected. But when these occurrences happen, they still show up as warnings and confuse users. This change demotes them to verbose messages.

Related work items: #1886476