Merged PR 632405: Productization of packed execution logging framework
This change completes and productizes the optimized BuildXL log analysis work originally started a year ago.

[Original deck](https://microsoft-my.sharepoint.com/:p:/p/rjelling/EeAmA2JdNehEqQBCzMdc8lsBpfFbTzJtpzDHaG5RHDU4zQ?e=KMDaYx)
[Original PR](https://dev.azure.com/mseng/Domino/_git/BuildXL.Internal/pullrequest/575957)

[Current deck with updated status](https://microsoft-my.sharepoint.com/:p:/p/rjelling/EQQrdyMUe4tAktNAs4AkaKcBCSNPE4LR0Gbly5TmERPY-w?e=hRPx1C)
[Recorded presentation to Windows devs](https://microsoft-my.sharepoint.com/:v:/p/rjelling/EQ8g2kCU5HVGge3GpuHjcF8BA5dEcpo6qxlg8B6B1jy2EA?email=rjelling%40microsoft.com)

This code has been tested with a full (non-cached) Windows build:

|Data|XLG analysis|PackedExecution|
|-----|--------------|-----------------|
|Uncompressed file sizes|22.4GB (21.9GB XLG + 0.5GB tables)|1.6GB|
|Compressed file sizes|9.2GB (8.7GB zipped XLG + 0.5GB zipped tables)|0.35GB (357MB)|
|Time to load and analyze after the build|46 minutes (dev workstation, FileConsumptionAnalyzer only)|26 seconds (dev workstation, all file consumption & pip execution data)|
|Duration of uncached all layers build|10 hrs 13 mins|10 hrs 12 mins (within normal variability, e.g. no visible overhead)|

The FileConsumptionAnalyzer results for this build were verified to be exactly consistent with the data from the PackedExecutionExporter. Details on request.

A viewer for this data format has been built, with a demo available here:
```
\\winbuilds\buildfile\PCV\BuildXL_Demo\1007_090757_demo_2.zip
```
Copy locally, unzip, run runme.cmd from Windows Explorer. This is all the data for a full all layers build. Will take about 8GB of memory and loads in 26 seconds on a 12 core dev box. Ping me if you cannot access this file, or if you have questions about the viewer or the data.

(A subsequent PR will hopefully bring the viewer code into the BuildXL repository as well, but it requires .NET Core 5 so I have to sort that out with Sergey.)

There is nothing Windows-specific about this work; in principle this viewer and this data set can be useful to and used by all BuildXL customers. The viewer is Windows Forms-based but the Forms dependency is minimal (just tree views and menus, basically).

I hope to enable this for all BuildXL builds of Windows before the end of October.

Related work items: #1858158