Merged PR 634795: Using C# Source Generators for generating logs in Bxl
The main work was done by @<Pasindu Gunasekara 🍣> and I was integrating the new approach into our sdk.

This PR adds `LogGenerator` project that provides C# `SourceGenerator` for creating logs.

The benefit of this approach is that we don't need another compiler invocation and separate dlls are not produced if the project has generated logs. Instead those logs are produced at one step. So it should reduce the size of the output directories and speed up the self host build.

We don't change the way we generate logs, instead the solution is based on the existing log gen logic but it is called by the comipler.

**Considerations**
To make the transition simpler we now support two ways of generating logs via `generateLogs` argument and `generateLogsInProc`. The first one is the old one and the second one is the new one.

Also there is an environment variable that allows switching the mode globally from the old one to the new one.