Merged PR 636904: Stop tracing the errors when GetCheckpointStateAsync times out
`AzureBlobStorageCheckpointRegistry.GetCheckpointStateAsync` has 30 seconds timeout (the default value) and when the timeout is hit, the cancellation token used inside that method is set.

This causes `CheckpointState.FromJsonStreamAsync` method to fail with `TaskCanceledException` that is later logged as error in the catch block.

This PR adds the logic to respect the cancellation to avoid the traces like this one:

![image.png](https://dev.azure.com/mseng/9ed2c125-1cd5-4a17-886b-9d267f3a5fab/_apis/git/repositories/50d331c7-ea65-45eb-833f-0303c6c2387e/pullRequests/636904/attachments/image.png)