Merged PR 636813: Wait for LLS initialization in UnregisterAsync call
In some cases the self check procedure detects that the file is incorrect and calls `UnregisterAsync` before LLS is fully initialized.

We had a similar issue when `QuotaKeeper` was trying to purge the content too early. To avoid this issue `DIstributedContentStore` has a function called `WaitForPostInitializationCompletionIfNeeded` that is called in `GetHashesInEvictionOrder` (this is the method that `QuotaKeeper` calls.

But a similar call was missing from `DistributedContentStore.UnregisterAsync` that can be called too early by the self checker.

This PR addresses this issue that should avoid having contract violation caused by this race condition.

Related work items: #1891052