Merged PR 636951: Avoid double tracing failures in FireAndForget
This PR changes the default behavior for `FireAndForget` method that takes `Task<BoolResult>` to not trace the error by default.

In many cases, the `Task<BoolResult>` is obtained from `PerformOperationAsync` call and is traced already. So we only need to trace the error when the task is obtained in a different way which is happening way more rarely.