Merged PR 628252: Only check HistoricPerfDataTable once per pip in ChooseWorkerCpu
Currently we are querying the HistoricPerfDataTable every time ChooseWorkerCpu is run: this may run multiple times per pip because it can be unsuccessful and retried later. Setting the RunnablePip's field to the `Empty` structure avoids this and keeps the hit/miss counters in the table consistent with the number of pips.