Merged PR 638098: Allow RuntimeCacheMissAnalyzer until the build ends to potentially load
Currently the RuntimeCacheMissAnalyzer is only allowed to load until it receives its first request. If it is not loaded by that point, the load task's result is ignored and runtime cache miss analysis doesn't kick in.

This change shelves all of the ExecutionLog events either until the RuntimeCacheMissAnalyzer loads or the build finishes (whichever is first). The tradeoff is memory at runtime. It does not address the bad pattern ignoring the result of the initialization task. Unfortunately, that operation isn't cancellable.