Merged PR 640113: Trace discarded cache settings as Debug message
Currently, during each startup, the cache service emits the following error log entry:

```
Named cache 'DefaultSSD' was discarded since the drive required by K:\dbs\Cache\ContentAddressableStore does not exist or is inaccessible on the machine.
```

It used to be a real problem if the cache configuration had a setting that corresponded to an unavailable drive. But we changed the design in a way that now this is a new norm and every production stamp now has 'DefaultSSD' configuration that targets the 'K' drive. This was done to simplify the configuration when different machine functions have different drive layouts (like the global cache machines do have 'K' drive, but command agents only have 'D' drive).

This PR changes the severity of the trace and makes it clear that this situation is a normal case that we do expect to happen all the time in production.