Merged PR 640224: Remove postconditions from the codebase
This is the last PR that cleans up the Code Contracts usage from the codebase.

Postconditions are not supported by 'Runtime Contracts' and technically can be used as the documentation, but because no one pays attention to them its likely that some of them already stale.

Instead of relying on potentially stale data, it would be better just to remove them. Plus the upcoming version of Runtime Contracts will warn if you try using them.