Merged PR 640219: Don't request SSL certificate from client
Grpc certificate is not needed from the client. The server certificate is enough to make the communication encrypted. Requesting the client certificate was more related to the authentication rather than encryption. I will start requesting the certificate from the client after we have a proper certificate verification method and .NET 6.0.

Related work items: #1898831