Merged PR 640117: Remove 'Pure' attribute usage
I'm working on 1.0 release of 'RuntimeContracts' that we use instead of Code Contracts and I'm planning to get rid of the attributes that doesn't make sense any more.

One of such attributes is `Pure`. Even though technically its not a bad attribute, our usage of this attribute in the code base is inconsistent and in some cases is not even correct (I removed a few cases when a method was marked with 'Pure' even though it was changing the instance state).

So this PR removes all the usages of this attribute from the codebase.

The PR has no semantic or runtime changes.