Merged PR 640887: Unconditionally Log DX code in DevOps listener
Unconditionally Log DX code in DevOps listener. Previously this was using the DevOps UI feature to specify file, line, column, and error code. But since we do not have that information for all events it was only sometimes being specified. Only specifying the code without the other fields will cause it to not be redendered.

The usefulness of repeating the file, line, and column is a bit dubious as well especially for generated specs. And those fields are still included in the pip description.

This change just makes the DX code part of the error message and removes the specification of those special fields. Example:
![image.png](https://mseng.visualstudio.com/9ed2c125-1cd5-4a17-886b-9d267f3a5fab/_apis/git/repositories/50d331c7-ea65-45eb-833f-0303c6c2387e/pullRequests/640887/attachments/image.png)

Related work items: #1899439