Merged PR 643162: Expose cache local only mode as a bxl command line argument
In order to disable remote cache on a per-build basis (context: JS prep build needs to opt out from the remote cache, whereas the JS main build should not) this PR exposes a bxl command line arg to only use the local cache.

To do that without roundtripping the option via the cache config file (in CB the same one is used for both builds) this change exposes the cache config object to the cache initialization method of ICacheFactory. The vertical aggregator then considers the configuration for disabling the remote cache.