Merged PR 643971: Simplify RunInSubst usage with /runInSubst
Facilitate subst usage by embedding in bxl the capability of running RunInSubst.exe with a newly introduced /runInSubst command line flag. This option addresses simple scenarios, where the subst target is always B:\ and the subst source defaults to the repo root (with an option to specify the latter explicitly, we have some repos where the main config file is not at the repo root).

Main target user is a developer using dev cache, where subst has to be run both on CB and locally in order to get cache hits.

Related work items: #1894881