Merged PR 643886: Remove the naming convention for 'async' methods in favour of the analyzers
There is a major limitation in enforcing naming convention for "async" methods: the editor config can enforce the naming only for the methods marked with 'async' keyword. I.e. with the existing configurations we had the following:

```csharp
// Fine: the method is marked with 'async' keyword
public async Task FooBarAsync() {await Task.Yield();}

// "Warning": Async suffix is redundant
public Task BazAsync() => FooBarAsync();
```

The .editorconfig-based implementation is very naive and it doesn't know the semantics of the code, so it can't enforce the same rule based on the result type.

On the other hand some analyzers, like https://www.nuget.org/packages/Microsoft.VisualStudio.Threading.Analyzers can enforce the naming convention properly.

So this PR just removes the .editorconfig option and the next PR will enable the same rule based on the analyzer.

Here is an example of a real false positive warning:
![image.png](https://dev.azure.com/mseng/9ed2c125-1cd5-4a17-886b-9d267f3a5fab/_apis/git/repositories/50d331c7-ea65-45eb-833f-0303c6c2387e/pullRequests/643886/attachments/image.png)