Merged PR 644747: When node.exe is specified, explicitly add its location to PATH of the graph construction process env
There is an implicit requirement coming from many JS coordinators (yarn/rush) to have node.exe in the PATH of the running process. PATH can be controlled by the user, so this is not necessarily the case. But when node.exe location is explicitly specified, make sure we expose it to PATH of the running graph construction tool.