Merged PR 645124: Retry async pipe reading on cancellation
In some builds, the pipe reading got cancelled in the middle of the builds. It is still unknown what caused the cancellation because none of BuildXL code (except the test code) performs cancellation on pipe reading. This PR tries to mitigate that issue by retrying the pipe reading when some unknown cancellation is requested. Since the cancellation is not requested by BuildXL itself, retrying pipe reading can increase reliability.

For now, this is an opt-in feature. QuickBuild may use this feature to handle issues in some of their customer build queues.