Merged PR 645561: Disable calculating setup cost and enable fireForgetMaterializeOutput by default
Calculating setup cost is expensive and it is not worth it. It was explicitly disabled for a lot of queues, now it is a good time to disable it by default in BuildXL.

We also enable fireForgetMaterializeOutput by default as it was explicitly enabled for all Office builds for a long time.