Merged PR 645254: ClusterState in Azure
This PR adds support for changing the underlying storage of ClusterState from Redis to Azure Blob Storage. It does so by abstracting away the logic that updates ClusterState from LLS to a ClusterStateManager. The ClusterStateManager has a IClusterStateStorage which can be either Redis, Azure Storage, or a Transitional layer used just for the transition.

When in Transitional mode, machines are force registered in Azure Storage with the same machine mappings as set by Redis, and all operations are also executed in storage -albeit results are ignored-. This paves the way for moving to the storage-only mode, in which we only read from storage.

Related changes included in this PR:

- Removing of marking machines as active in MachineReputation and Event processing. The reason here is that doing so may cause incorrect decisions to be taken. For example, an completed copy cause machine A to report a good reputation on a machine B which has declared itself as inactive; this would lead machine A to believe B is up even when it isn't, which could cause a build failure if A made a decision to pin based on B's existence.
- Stop metadata service clients from reading the master machine location from ClusterState. Instead, they now read directly from the leader election mechanism. This both helps prevent staleness and simplifies the code.
- If Azure Storage Emulator is running locally, you're able to (manually) run all tests against it and they pass

Related work items: #1906346