Merged PR 646175: ModuleAffinityLoadFactor for experimentation
When choosing a worker for a module, we need to fill the 2x capacity of the preferred worker before the next workers. Running a module in another worker can be expensive. That's why, we try to fill the double capacity for the first preferred worker. We did not experiment with the load factor for module affinity, so I want to do some abtesting to find a better load factor. We should find a sweet spot between end-to-end duration and the size of the materialized inputs.