Merged PR 647053: Don't snap ProcessRetries setting into serialized pips
ProcessBuilder only executes as part of graph construction. It cannot utilize data settable on the command line since the pip graph won't be recreated when that command line changes. Otherwise, it won't be honoring the command line of the current build.

This is the only instance where ProcessBuilder is consuming the configuration object so this exact category of issue looks to be a one off. However if configuration impacting how pips are run is used in other ways in DScript logicm, similar more subtle issues may exist.

Related work items: #1912548