Merged PR 644503: Add channel-based ActionBlockSlim implementation
This PR adds an optional implementation of `ActionBlockSlim` that is based on `System.Threading.Channels.Channel` instead of a manual semaphore-based implemtation.

The new implementation is optional and should be enabled globally.

I did performance testing to compare the two version. And the channel-based version was faster (like 20-sh percents) in all the scenarios.

Related work items: #1888888