Merged PR 648198: Log MaterializeOutputOverhang duration
The critical path logging for metabuild can be confusing as we do not include materializeoutput duration in the critical path. It got even more confused with FireForgetMaterializeOutput feature which is enabled by default.

I decided to log the materialize overhang duration which means how long it takes to send the materializeoutput requests to all workers after scheduler is done with everything except materializeoutput.

Related work items: #1848901