Merged PR 648435: Use queue to store outputs in the cache
For pips with a high number of outputs, the current approach that spawns as many tasks as outputs to store incurs in significant lock contention (profiler shows between 30% and 60% of the time spent on lock contention when storing a pip with ~1M outputs). This PR uses a queue instead in order to get the number of spawned tasks under control. Experiments show a 15%-20% time improvement for the aforementioned pip.

@<Sergey Tepliakov> , the channel-based queue showed a slight improvement over the regular ActionBlock queue (with some variance that doesn't allow me to be 100% sure about it). But overall it is performing equally or better, so I decided to use it.