Merged PR 650385: Experimental asynchronous pipe reader
As part of our effort in dealing with pipe reading issue in .NET6, we implemented 2 async pipe reader, StreamReader-based and Pipeline-based. The former is recommended by the .NET team. However, our microbenchmarking showed that StreamReader-based one is twice slower than the Pipeline-based.

We'll do ab-testing on these pipe readers to know if they address the pipe reading issue, and then to know their perf. characteristics.