Merged PR 651684: Enable server mode when remoting is enabled
Previously, the server mode wasn't compatible with process remoting. The root cause of the issue is because we use the singleton pattern to get-or-create process remoting manager. Thus, that manager is used across build sessions, but was actually disposed after the very first session.

This PR stops the use of singleton pattern, and simply creates a remoting manager for each build session. That creation happens inside the scheduler, and the obtained instance will be disposed when the scheduler is disposed.

Related work items: #1926109