Merged PR 652073: Do not shut down components that were not started
This PR changes the shutdown logic in `StartupShutdownCompoentBase` to avoid runtime failures when the nested component is shut down but it was never started up.

This PR fixes the following errors that are happening in prod when the startup of one of the component fails:

```
Critical error occurred: GrpcConnectionPool.ShutdownAsync stop 205.883ms. result=[Error=[NullReferenceException: Object reference not set to an instance of an object.], Exception=[System.NullReferenceException: Object reference not set to an instance of an object.
   at Task<BoolResult> BuildXL.Cache.ContentStore.Distributed.MetadataService.GrpcConnectionPool.ShutdownCoreAsync(OperationContext context) in \.\Public\Src\Cache\ContentStore\Distributed\MetadataService\GrpcClientPool.cs:line 80
```

or

```
Critical error occurred: CheckpointManager.ShutdownAsync stop 3.1428ms. result=[Error=[ContractException: Assertion failed: Cannot shut down 'ContentLocationDatabase' because ShutdownAsync method was already called on the instance with Id=35..
	at \.\Public\Src\Cache\ContentStore\Library\Utils\StartupShutdownSlimBase.cs:126], Exception=[System.Diagnostics.ContractsLight.ContractException: Assertion failed: Cannot shut down 'ContentLocationDatabase' because ShutdownAsync method was already called on the instance with Id=35..
```